+++ apps/code-examples/src/app/code-examples/lists/filter/inline/filter-demo.module.ts
@@ -1,7 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { SkyIdModule } from '@skyux/core';
 import { SkyCheckboxModule, SkyInputBoxModule } from '@skyux/forms';
 import { SkyToolbarModule } from '@skyux/layout';
 import { SkyFilterModule, SkyRepeaterModule } from '@skyux/lists';
@@ -13,7 +12,6 @@ import { FilterDemoComponent } from './filter-demo.component';
     CommonModule,
     FormsModule,
     SkyCheckboxModule,
-    SkyIdModule,
     SkyInputBoxModule,
     SkyFilterModule,
     SkyRepeaterModule,