+++ apps/code-examples/src/app/code-examples/text-editor/text-editor-inline-help/text-editor-demo.component.ts
@@ -19,8 +19,4 @@ export class TextEditorDemoComponent {
       myText: new UntypedFormControl(this.richText),
     });
   }
-
-  public onActionClick(): void {
-    alert('Help inline button clicked!');
-  }
 }