+++ apps/code-examples/src/app/code-examples/validation/url-validation/control-validator/url-validation-demo.module.ts
@@ -1,7 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { SkyIdModule } from '@skyux/core';
 import { SkyInputBoxModule } from '@skyux/forms';
 import { SkyStatusIndicatorModule } from '@skyux/indicators';
 import { SkyUrlValidationModule } from '@skyux/validation';
@@ -13,7 +12,6 @@ import { UrlValidationDemoComponent } from './url-validation-demo.component';
     CommonModule,
     FormsModule,
     ReactiveFormsModule,
-    SkyIdModule,
     SkyInputBoxModule,
     SkyStatusIndicatorModule,
     SkyUrlValidationModule,