+++ apps/e2e/angular-tree-component-storybook/src/app/angular-tree-component/angular-tree-component.module.ts
@@ -1,13 +1,24 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { TreeModule } from '@circlon/angular-tree-component';
+import { SkyAngularTreeModule } from '@skyux/angular-tree-component';
+import { SkyHelpInlineModule } from '@skyux/indicators';
+import { SkyDropdownModule } from '@skyux/popovers';
 
 import { AngularTreeComponentComponent } from './angular-tree-component.component';
 
 const routes: Routes = [{ path: '', component: AngularTreeComponentComponent }];
 @NgModule({
   declarations: [AngularTreeComponentComponent],
-  imports: [CommonModule, RouterModule.forChild(routes)],
+  imports: [
+    CommonModule,
+    RouterModule.forChild(routes),
+    SkyAngularTreeModule,
+    SkyHelpInlineModule,
+    SkyDropdownModule,
+    TreeModule,
+  ],
   exports: [AngularTreeComponentComponent],
 })
 export class AngularTreeComponentModule {}