+++ apps/playground/src/app/components/layout/text-expand-repeater/text-expand-repeater-routing.module.ts
@@ -1,0 +1,26 @@
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+
+import { ComponentRouteInfo } from '../../../shared/component-info/component-route-info';
+
+import { TextExpandRepeaterComponent } from './text-expand-repeater.component';
+
+const routes: ComponentRouteInfo[] = [
+  {
+    path: '',
+    component: TextExpandRepeaterComponent,
+    data: {
+      name: 'Text expand repeater',
+      icon: 'bars',
+      library: 'layout',
+    },
+  },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class TextExpandRepeaterRoutingModule {
+  public static routes = routes;
+}