+++ apps/playground/src/app/components/layout/text-expand-repeater/text-expand-repeater.component.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-text-expand-repeater',
+  templateUrl: './text-expand-repeater.component.html',
+})
+export class TextExpandRepeaterComponent {
+  public repeaterData: string[] = [
+    'Repeater item 1',
+    'Repeater item 2',
+    'Repeater item 3',
+    'Repeater item 4',
+    'Repeater item 5',
+  ];
+}